pyproject.toml
[tool.poetry]
name = "rapidata"
version = "2.37.0"
description = "Rapidata package containing the Rapidata Python Client to interact with the Rapidata Web API in an easy way."
authors = ["Rapidata AG <info@rapidata.ai>"]
license = "Apache-2.0"
readme = "README.md"
packages = [{ include = "rapidata", from = "src" }]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
pyjwt = "^2.9.0"
pillow = "^10.4.0"
pydantic = "^2.8.2"
python-dateutil = "^2.9.0.post0"
tqdm = "^4.66.5"
deprecated = "^1.2.14"
colorama = "0.4.6"
tinytag = "^2.0.0"
pandas = "^2.2.3"
authlib = "^1.5.1"
httpx = "^0.28.1"
packaging = ">=16.0.0"
opentelemetry-api = ">=1.27.0"
opentelemetry-sdk = ">=1.27.0"
opentelemetry-exporter-otlp = ">=1.27.0"

[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.1"
black = "^24.8.0"

[tool.poetry.group.docs.dependencies]
mkdocs-material = "^9.5.34"
mkdocstrings = { extras = ["python"], version = "^0.26.0" }
mkdocs-autorefs = "^1.2.0"
mkdocs-gen-files = "^0.5.0"
mkdocs-literate-nav = "^0.6.1"
pygments = "^2.18.0"
mkdocs-section-index = "^0.3.9"
mike = "^2.1.3"
mkdocs-redirects = "^1.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
